+++ CHANGELOG.md
@@ -6,6 +6,18 @@ Most recent releases are shown at the top. Each release shows:
 - **Changed**: Additional parameters, changes to inputs or outputs, etc
 - **Fixed**: Bug fixes that don't change documented behaviour
 
+## 0.6.1 (2024-12-4)
+
+### new:
+- N/A
+
+### changed
+- Changed `pdf_use_unstructured` to `pdf_unstructured` and
+  `pdf2md` to `pdf_markdown` (#102)
+
+### fixed:
+- N/A
+
 ## 0.6.0 (2024-12-3)
 
 ### new: